Update contact list GUI
