+++ src/main/resources/view/ContactList.fxml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.layout.StackPane?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.VBox?>
+
+<HBox id="contactList" fx:id="contactListPane" styleClass="pane-with-border" minWidth="340" prefWidth="340"
+    xmlns="http://javafx.com/javafx/17" xmlns:fx="http://javafx.com/fxml/1">
+    <!-- Contact list -->
+    <VBox fx:id="personList" styleClass="pane-with-border" VBox.vgrow="ALWAYS" minWidth= "400" prefWidth="400">
+        <padding>
+        <Insets top="10" right="10" bottom="10" left="10" />
+        </padding>
+        <StackPane fx:id="personListPanelPlaceholder" VBox.vgrow="ALWAYS"/>
+    </VBox>
+    <!-- Contact details panel -->
+    <VBox id="personDetails">
+        <Label text = "Contact Details Here"/>
+    </VBox>
+</HBox>
