+++ src/main/resources/view/PersonListCard.fxml
@@ -32,8 +32,6 @@
         <Label fx:id="phone" styleClass="cell_small_label" text="\$phone" />
       </HBox>
       <FlowPane fx:id="tags" />
-      <!-- <Label fx:id="address" styleClass="cell_small_label" text="\$address" />
-      <Label fx:id="email" styleClass="cell_small_label" text="\$email" /> -->
     </VBox>
   </GridPane>
 </HBox>
